Refactoring vulnerability view to split header in the separate file
