+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/basic/organizations/OrganizationVulnerabilitiesTab.kt
@@ -31,7 +31,7 @@ private val vulnerabilityTable: FC<TableProps<VulnerabilityDto>> = tableComponen
                 Fragment.create {
                     td {
                         Link {
-                            to = "/${FrontendRoutes.VULN}/${cellContext.row.original.name}"
+                            to = "/${FrontendRoutes.VULNERABILITY_SINGLE}/${cellContext.row.original.name}"
                             +cellContext.value
                         }
                     }
